Fix codecheck

+++ mindspore_rl/utils/mcts/cpu/cpu_mcts_ops.cc
@@ -332,7 +332,7 @@ extern "C" int UpdateLeafNodeOutcome(int nparam, void **params, int *ndims,
   if (tree == nullptr) {
     return kErrorCode;
   }
-  int index = tree->visited_path().size() - 1;
+  int index = static_cast<int>(tree->visited_path().size() - 1);
   bool ret = tree->UpdateOutcome(return_value, index);
   tree->Memcpy(output, &ret, sizeof(bool));
   return 0;
@@ -375,7 +375,7 @@ extern "C" int UpdateLeafNodeTerminal(int nparam, void **params, int *ndims,
   if (tree == nullptr) {
     return kErrorCode;
   }
-  int index = tree->visited_path().size() - 1;
+  int index = static_cast<int>(tree->visited_path().size() - 1);
   bool ret = tree->UpdateTerminal(*terminal, index);
   tree->Memcpy(output, &ret, sizeof(bool));
   return 0;
@@ -417,7 +417,7 @@ extern "C" int UpdateLeafNodeState(int nparam, void **params, int *ndims,
   if (tree == nullptr) {
     return kErrorCode;
   }
-  int index = tree->visited_path().size() - 1;
+  int index = static_cast<int>(tree->visited_path().size() - 1);
   bool ret = tree->UpdateState(state, index);
   tree->Memcpy(output, &ret, sizeof(bool));
   return 0;
@@ -495,9 +495,9 @@ extern "C" int GetLastState(int nparam, void **params, int *ndims,
   if (tree == nullptr) {
     return kErrorCode;
   }
-  int index = tree->visited_path().size() - 2;
+  int index = static_cast<int>(tree->visited_path().size() - 2);
   if (index < 0) {
-    index += tree->visited_path().size();
+    index += static_cast<int>(tree->visited_path().size());
   }
   auto output_state = tree->GetState(index);
   tree->Memcpy(output, output_state, tree->state_size() * sizeof(float));
